pyproject.toml
[tool.poetry]
name = "pig_db"
version = "0.1.0"
description = ""
authors = ["“chie-hirabayashi” <“nyantoroman@yahoo.co.jp”>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
peewee = "^3.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
